Double to Float

+++ stream_context/src/main/scala/ConsumeKafka.scala
@@ -32,7 +32,7 @@ object ConsumeKafka {
         val lines = rdd.map(_._2)
         val ticksDF = lines.map( x => {
                                   val tokens = x.split(",")
-                                  Tick(tokens(0).toInt, tokens(1), tokens(2), tokens(3), tokens(4).toDouble)}).toDF()
+                                  Tick(tokens(0).toInt, tokens(1), tokens(2), tokens(3), tokens(4).toFloat)}).toDF()
         val ticks_per_source_DF = ticksDF.groupBy("source_id")
                                .agg("quantity" -> "avg", "quantity" -> "sum")
                                .orderBy("source_id")
